Create Synta_PIC_Optimized.pde
Code base on original AstroEQ firmware. some stuff like scalar and rate calculation refinement are removed to simplify the code and to be as efficient as possible. Synta commands are not in a separate library since it also add some efficiency to the probale detriment of code lisibility.
Hardware wiring ideas will be published in a next commit