Massive tidy up of AstroEQ cpp file
Along with tidying up the tab size (4 spaces rather than 2), commenting
has been improved as has readability.

(1) New #defines for setting pin direction, value and reading pins have
all been made. These make code which changes pins much clearer.

(2) The arrangement of the functions has been tidied up so they are
better grouped together. Sections have been named for things like global
variables, etc.

(3) Support has been added for the DRV8834 though this is still
preliminary. This has required a rewrite of the mode mapping function.
Rather than using an array of values, a bit mask is used for the mode
pins instead which makes generating the mapping and accessing the values
much cleaner.

(4) Remove the decelerationSteps requirement in the motor routines. This
was added to ensure go-to accuracy but actually was pointless because
the way the deceleration steps were chosen will result in the correct
go-to length without this. Removing this also removes the abusive use of
the PORTC and EEPROM registers as high speed IO registers which make the
firmware a little less /weird/. It also means that PORTC I/O pins are
actually no longer reserved.

(5) Convert Goto control/status flags to use a bitmask in a single IO
register rather than variables in the SRAM. This massively optimises
checking of flags as simple SBIC instructions can be generated by the
compiler rather than LDS, ORI, STS, etc. It also saves a bit of SRAM not
that that really matters. Inline functions have been made to neaten up
access to these flags.

(6) Add support for a new (as yet untested) standalone mode which will
make use of two GPIO pins and the ST4 port to allow manual control when
EQMOD is not connected. Two speeds as well as sidereal tracking will be
supported - 2x for fine movements and Go-To speed for fast movements
(the speed programmed by the config utility). The speed will be
controlled by one GPIO pin and entry to standalone mode controlled by a
second.

(7) Rewrite ST-4 ISR code to work with standalone mode. Also simplify
the ISR a bit by storing precalculated ST-4 speeds in variables rather
than calculating it using optimised assembler code on the fly.

(8) Reduce complexity of motor ISRs. Before the code was written so goto
movements reach the deceleration point and then adding it on to the
target before continuing the goto during deceleration to that point
(quite convoluted). This has been simplified so that once the
deceleration point has been reached, it will switch into slew mode
(which saves some calculation) and decelerate to stopped. This may in
practice result in goto movements being one or two steps short, but that
doesn't really matter as there is usually a short correction movement
issued by EQMOD to account for RA movement, so in practice this should
never affect accuracy. What it does do is reduce the amount of
calculation which reduces the likelihood of something crashing.