Major Firmware Changes
There have been many changes since the last commit, comprising of
tidying of the source code, completion of the standalone handcontroller,
and major changes to the motor routines.

1. Firstly some further tidying and rearranging of code to neaten things
up.

2. The standalone handcontroller routines have recieved some updates.
The standalone mode detection routines along with serial communications
for the advanced mode have been compeleted. The advanced mode requires
an external resistor not present in hardware versions pre V4.6, so the
detection of advanced controllers can be disabled through the config
utility.

3. The motor driver routines have been rewritten. There are two reasons
for this. The first is that the routines are now pure C - no longer are
they littered with assembly code. Additionally and more importantly, the
acceleraiton and deceleration routines have been changed to now be
controlled via lookup table which allows for a linear accel/decel of the
mount which was previously very jerky.