New AstroEQ Hardware Version (V4.7)
This new version is designed for machine assembly by using a mixture of
SMD and PTH components. The board is functionally identical to V4.6
which will remain the version being used for blank PCBs. I recommend
anyone doing DIY to use the V4.6 hardware.

Additionally I have updating my SMD EQ5 design to the V4.7 hardware
version. It was previously a V4.0 design. The upgrade is purely because
I want to be able to test the latest hardware with my mount which uses
this design.