Major Firmware Improvements
1. Tidy up commands.h and commands.cpp to correct some indentation and
improve readability.
2. Add new ST-4 speed constant to firmware. This allows ST4 factor to be
changed through the configuration utility
3. Rename st4Pin to st4Pins as it is an array of all ST4 pins.
4. Rewrite whole ST-4 processing. Remove the pin change interrupt that
was driving ST-4 and instead use a polling approach in the main()
function. This allows for better resiliance to bounce of the inputs and
prevents the controller from missing the end of an ST4 pulse and not
stopping.
5. Bump version number in code which was still saying 7.5
6. Complete basic handcontroller mode, including a new routine to detect
the required speed (three speeds are possible, ST4 speed, 2x speed and
Goto speed.
7. Rearrance command decoder switch to seperate out programming commands
into a different structure. This prevents them accidentally being
executed if the mount is in run mode.