Update AstroEQ Atmel Studio Solution/Project Files
Add project files for the Arduino Mega 2560 and 1280. These use the same
source code, but seperate projects are needed to ensure the correct MCU
is set and the custom programming tool strings are correct.

Additionally a post build even has been added to all projects which will
automatically copy the release firmware file into the AstroEQ Config
Utility Hex files folder in the main repository. It will also update the
version number in the versions.txt file based on the VerNum.txt file in
the firmware folder.