Update Displayed Firmware Version When Fetching New Firmware
The firmware version shown by the drop down menu was not being updated
after a new firmware version is downloaded. This is fixed now. For
longer version numbers (e.g. 8.03) the number was getting cut off, so
more space has been made available for the number to be displayed in.