Fix Atmel Studio Solution Toolchain Settings
The toolchain was mistakenly set to include the '-funsigned-char' flag
which was causing issues with the Synta parser when it came to handling
errors. This in turn meant that it was not possible to correctly
initialise the EEPROM (it would always report as initialised to the
config utility) and so broke EQMOD support for brand new controllers.