Make AstroEQ.h Load Version Number from VerNum.txt
VerNum.txt is a text file in the AstroEQ Atmel Studio project which
contains the current version number. This value is used by build event
scripts to update the firmware version in the config utility version
records.

Using a bit of clever preprocessor trickery, the VerNum.txt file is
loaded in to the code at compile time and converted into an inline
function which optimises to a constant value of that version number,
meaning bumping the version number is now a simple case of changing the
text file and recompiling.