Fix Atmel Studio Solution Settings
The solution settings for the AstroEQ project were wrong which was
resulting in every version of the firmware being compiled for ATMega162
- even the Arduino Mega projects.
To solve this each project in the solution requires its own
configuration name to keep intermediate files from getting shared
between different processors.

This is the root cause of issue #20.