Setup AVRdude as Custom Programming Tool
The previous attempt at using AVRdude as a custom programming tool
worked but would always say that there was an error even when everything
was successful. The reason is because AVRdude uses STDERR for its
verbose output, not STDOUT.

The fix is to use a new batch file to redirect the STDERR output of
AVRdude to STDOUT to allow correct handling by Atmel Studio 7.