Patch Config Utility to Use "avrdude" Package (Fixes #22)
This is a fix to issue #22 in which the distributed i386 version of
avrdude was not functional on other platforms such as amd64 or armhf.
This patch changes the directories of the avrdude files in linux to
point to those included with the avrdude package which will already be
of the correct architecture to match os.