Rebuilt Config Utility for Linux to Fix #22
Rebuilt the config utility to include fix for issue #22. The changes
also update the dependencies of the utility to the correct version of
JSSC (>=2.8.0) and to include dependence on the avrdude package.

The fix also migrates the used JSSC java files to be those included with
libjssc-java rather than jusing the java file that Processing generates
with the natives from that package. This change fixes another issue
whereby on some systems a "java.lang.UnsatisfiedLinkError:
jssc.SerialNativeInterface.openPort" exception was occuring.