Correct handling of Mount Status :f command
Errm, yeah. Somehow for the last 5 years the AstroEQ has worked with
this error....

Bascially the bits in the :f response were muddled up so that it was
reporting "stopped/!running" instead of "slew/!goto", and "goto/!slew"
instead of "running/!stopped". Because of the inversion of the bits as
well as them being swapped, EQMOD managed to cope OK it seems. INDIlib
on the other hand was getting upset and doing strange things (and
rightfully so!)...

Anyway, this major WTF moment has been resolved. I've also added in a
missing highspeed/lowspeed flag to help INDI understand what is going on
internally. EQMOD doesn't use this flag, but no harm it being there.