Improve :O2x Handling
The purpose of EQMOD using the :Oxx command has become clearer. This is used to control up to two GPIO pins, for example to trigger a camera shutter.

The :O1x command in AstroEQ is used to enter programming mode, so in order to tell EQMOD not to use the SNAP1 interface, AstroEQ now reports an error packet when :O1x is issued in run-mode. The configuration utility knows to expect this error packet. On the 20th time of writing :O1x, AstroEQ enters programming mode and sends a success packet.

This should provide a secure way of fixing issue #24.

AstroEQ V4.6 and newer, as well as Mega variants, have a spare GPIO pin not used for standalone mode. To still allow a feature rich experience in AstroEQ, the :O2x command is now used to control this GPIO pin.

By adding a simple transistor or optocoupler, the GPIO1 pin can now be used as SNAP2 in EQMOD.