All Different cmd Payload/Response Lengths in Prog
So the command set has been extended in the skywatcher documentation, which is now causing problems because I use some of the commands for my own purpose in programming mode. For now it is just the 'q' command which requires a different payload length in the protocol than I use.

To solve this, we now have the ability to set different lengths for programming mode and run mode commands by using the upper nibble of the payload length parameter which was always zero before.